Reduced some compile times using .cpp files and precompiled Headers for the operation.
