+++ src/engine/GroupBy.h
@@ -12,6 +12,7 @@
 #include <utility>
 #include <vector>
 
+#include "gtest/gtest.h"
 #include "../parser/Alias.h"
 #include "../parser/ParsedQuery.h"
 #include "./Operation.h"
