+++ src/parser/data/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_library(parserData Iri.cpp BlankNode.cpp Variable.cpp)
