+++ src/parser/data/Iri.h
@@ -7,16 +7,12 @@
 #include <string>
 
 #include "./Context.h"
-#include "ctre/ctre.h"
 
 class Iri {
   std::string _string;
 
  public:
-  explicit Iri(std::string str) : _string{std::move(str)} {
-    AD_CHECK(ctre::match<"(?:@[a-zA-Z]+(?:-(?:[a-zA-Z]|\\d)+)*@)?"
-                         "<[^<>\"{}|^\\\\`\\0- ]*>">(_string));
-  }
+  explicit Iri(std::string str);
 
   // ___________________________________________________________________________
   // Used for testing
