+++ src/parser/data/Variable.cpp
@@ -1,0 +1,18 @@
+//  Copyright 2022, University of Freiburg,
+//                  Chair of Algorithms and Data Structures.
+//  Author: Johannes Kalmbach <kalmbach@cs.uni-freiburg.de>
+
+#include "parser/data/Variable.h"
+
+#include "ctre/ctre.h"
+
+// ___________________________________________________________________________
+Variable::Variable(std::string name) : _name{std::move(name)} {
+// verify variable name starts with ? or $ and continues without any
+// special characters. This is weaker than the SPARQL grammar,
+// but it is close enough so that it will likely never cause issues.
+AD_CHECK(ctre::match<"[$?]\\w+">(_name));
+// normalise notation for consistency
+_name[0] = '?';
+}
+
