+++ src/parser/data/Variable.h
@@ -8,7 +8,6 @@
 #include <utility>
 
 #include "Context.h"
-#include "ctre/ctre.h"
 #include "engine/ResultTable.h"
 #include "index/Index.h"
 
@@ -16,17 +15,13 @@ class Variable {
   std::string _name;
 
  public:
-  explicit Variable(std::string name) : _name{std::move(name)} {
-    // verify variable name starts with ? or $ and continues without any
-    // special characters. This is weaker than the SPARQL grammar,
-    // but it is close enough so that it will likely never cause issues.
-    AD_CHECK(ctre::match<"[$?]\\w+">(_name));
-    // normalise notation for consistency
-    _name[0] = '?';
-  }
+  explicit Variable(std::string name);
 
   // Todo<joka921> There are several similar variants of this function across
   // the codebase. Unify them!
+  // TODO<joka921> This function can also be in the .cpp file, but we first
+  // have to figure out the link order.
+
   // ___________________________________________________________________________
   [[nodiscard]] std::optional<std::string> evaluate(
       const Context& context, [[maybe_unused]] ContextRole role) const {
